+++ packages/expo-modules-core/ios/Core/Views/SwiftUI/SwiftUIHostingView.swift
@@ -7,6 +7,7 @@ import SwiftUI
  */
 internal protocol AnyExpoSwiftUIHostingView {
   func updateProps(_ rawProps: [String: Any])
+  func getContentView() -> any ExpoSwiftUI.View
 }
 
 extension ExpoSwiftUI {
@@ -21,6 +22,7 @@ extension ExpoSwiftUI {
      It's an environment object that is observed by the content view.
      */
     private let props: Props
+    private let contentView: any ExpoSwiftUI.View
 
     /**
      View controller that embeds the content view into the UIKit view hierarchy.
@@ -31,8 +33,8 @@ extension ExpoSwiftUI {
      Initializes a SwiftUI hosting view with the given SwiftUI view type.
      */
     init(viewType: ContentView.Type, props: Props, appContext: AppContext) {
-      let rootView = ContentView().environmentObject(props)
-
+      self.contentView = ContentView()
+      let rootView = AnyView(contentView.environmentObject(props))
       self.props = props
       self.hostingController = UIHostingController(rootView: rootView)
 
@@ -65,6 +67,13 @@ extension ExpoSwiftUI {
       } catch let error {
         log.error("Updating props for \(ContentView.self) has failed: \(error.localizedDescription)")
       }
+    }
+
+    /**
+     Returns inner SwiftUI view.
+     */
+    public func getContentView() -> any ExpoSwiftUI.View {
+      return contentView
     }
 
     /**